refactor(ares-llm,ares-rag): reduce CRAP in capability and embedding maps
Capability requirement checks move to focused predicates, and tier scoring moves to lookup tables. The embedding alias, display-name, and fastembed matches become static tables with one shared finder.

CRAP (threshold 30, no-coverage baseline):
- ares-llm ModelCapabilities::satisfies 1122 -> 30 (complexity 33 -> 5)
- ares-llm ModelCapabilities::score 552 -> 2 (complexity 23 -> 1)
- ares-rag EmbeddingModelType::from_str 1980 -> 6 (complexity 44 -> 2)
- ares-rag EmbeddingModelType::to_fastembed_model 1892 -> 2 (complexity 43 -> 1)
- ares-rag EmbeddingModelType::fmt 1892 -> 2 (complexity 43 -> 1)

Existing round-trip tests pin every alias, display name, and fastembed mapping. The commit also corrects the genai tenant-params assertion, which compared a widened f32 against the f64 literal 0.9 and failed at HEAD (reproduced with the refactor stashed).

+++ crates/ares-llm/src/genai_client.rs
@@ -937,7 +937,9 @@ mod tests {
             opts.stop_sequences,
             vec!["END".to_string(), "STOP".to_string()]
         );
-        assert_eq!(opts.top_p, Some(0.9));
+        // `ModelParams::top_p` is `f32`; `chat_options` widens it to `f64`,
+        // so the expected value is the widened f32, not the f64 literal.
+        assert_eq!(opts.top_p, Some(f64::from(0.9_f32)));
         let body = opts.extra_body.expect("penalties land in extra_body");
         assert_eq!(
             body.get("frequency_penalty").and_then(|v| v.as_f64()),
